EES-1716 - moved test cases from one class to its duplicate from other branch
